Sam batch normalisation

+++ src/utils/helpers.py
@@ -44,7 +44,7 @@ def create_loss_function(loss_name: str):
 
 
 def create_optimiser(
-    optimiser_name: str, layers: list, lr: int, weight_decay: float, momentum: float
+    optimiser_name: str, layers: list, lr: float, weight_decay: float, momentum: float
 ):
     """
     Creates a optimiser and returns it
